Tests/LibWeb/Text/input/HTML/Window-prototype.html

<script src="../include.js"></script>
<script>
    test(() => {
        println(window)
        println(window.__proto__)
        println(window.__proto__.__proto__)
        println(window.__proto__.__proto__.__proto__);
        println(window.__proto__.__proto__.__proto__.__proto__);
        println(window.__proto__.__proto__.__proto__.__proto__.__proto__);
    });
</script>
