Tests/LibWeb/Text/expected/HTML/Window-prototype.txt

[object Window]
[object Window]
[object WindowProperties]
[object EventTarget]
[object Object]
null
